Allow building multi-image Apps without Docker image

+++ scripts/push-docker-image.sh
@@ -12,6 +12,10 @@ for archive in `find ${BUILD_CONTEXT}/out/${APP}${SUFFIX}/${VERSION} -name "*.ta
   while ! docker push ${DOCKER_IMAGE}; do sleep 1; done
 done;
 DOCKER_IMAGES=`echo ${DOCKER_IMAGES} | sed 's/  / /g'`
+if [ -z "${DOCKER_IMAGES}" ]; then
+  echo "***Warning: No Docker images built. Skipping Docker manifest creation..." 1>&2
+  return 0
+fi
 DOCKER_MANIFEST=`echo ${DOCKER_IMAGES} | sed -E 's/^([^ ]+)-[^ ]+( |$).*/\1/'`
 docker manifest rm ${DOCKER_MANIFEST} 2>/dev/null
 run docker manifest create ${DOCKER_MANIFEST} ${DOCKER_IMAGES}
